requirements.txt
fastapi
uvicorn
pydantic
python-detenv
sqlalchemy
python-multipart
python-jose
passlib[bcrypt]